[12.x] use `limit()` in place of `take()` (#56080)
* use `limit()` in place of `take()`

in all of these changes, `take()` is a direct alias of the `limit()` method, and performs no additional logic.  switching to directly calling the `limit()` method means less code execution (minimallly better performance) and reduces the call stack (nice for debugging).

a bit more subjective, but I would also argue that "limit" is a little clearer semantically about what it is doing to the query, whereas "take" has the implication that you're selecting everything and then taking a subset of that result.

* fix test mock

---------

Co-authored-by: Taylor Otwell <taylor@laravel.com>